jobartifacts: follow-on updates
per comments:
* make json attrs snake-cased
* move testing utils to their own file

also:
* minor simplifications to manager event loops
