+++ pkg/util/utils.go
@@ -1,22 +1,14 @@
 package util
 
 import (
-	"context"
 	"fmt"
 	"math"
 	gourl "net/url"
-	"os"
 	"regexp"
 	"strconv"
-	"testing"
 	"time"
 
-	"cloud.google.com/go/storage"
 	v1 "github.com/openshift/sippy/pkg/apis/sippy/v1"
-	"github.com/openshift/sippy/pkg/dataloader/prowloader/gcs"
-	"github.com/openshift/sippy/pkg/db"
-	"github.com/sirupsen/logrus"
-	"gorm.io/gorm/logger"
 )
 
 type FailureGroupStats struct {
@@ -184,48 +176,4 @@ func AdjustReleaseTime(relTime time.Time, isStart bool, daysAdjustment string, c
 		}
 	}
 	return relTime
-}
-
-/*
-	  Enable functional tests requiring a live database and/or GCS bucket with known data to run,
-	  but we do not want them trying to run during CI, so skip tests whose required environment variables are not set.
-	  Don't risk checking in credentials with code; supply them with environment variables:
-		TEST_DB_LOG_LEVEL: "silent" or "info" or "warn" or "error" - the log level for gorm database methods
-		TEST_SIPPY_DATABASE_DSN: the DSN for the sippy postgres database e.g. postgresql://sippyro:...@sippy-postgresql...amazonaws.com/sippy_openshift
-		TEST_GCS_CREDS_PATH: the path to a local GCS credentials file, e.g. /home/$USER/git/sippy/openshift-sippy-ro.creds.json
-*/
-const GcsBucketRoot = "test-platform-results"
-
-func GetDbHandle(t *testing.T) *db.DB {
-	dbLogLevel := os.Getenv("TEST_DB_LOG_LEVEL") // e.g. "info" or "silent"
-	if dbLogLevel == "" {
-		dbLogLevel = "silent"
-	}
-	gormLogLevel, err := db.ParseGormLogLevel(dbLogLevel)
-	if err != nil {
-		logrus.WithError(err).Errorf("Cannot parse TEST_DB_LOG_LEVEL %s", dbLogLevel)
-		gormLogLevel = logger.Silent
-	}
-
-	dsn := os.Getenv("TEST_SIPPY_DATABASE_DSN")
-	if dsn == "" {
-		t.Skip("TEST_SIPPY_DATABASE_DSN environment variable is not set; skipping database tests")
-	}
-	dbc, err := db.New(dsn, gormLogLevel)
-	if err != nil {
-		logrus.WithError(err).Fatal("Cannot connect to database")
-	}
-	return dbc
-}
-
-func GetGcsBucket(t *testing.T) *storage.BucketHandle {
-	pathToGcsCredentials := os.Getenv("TEST_GCS_CREDS_PATH")
-	if pathToGcsCredentials == "" {
-		t.Skip("TEST_GCS_CREDS_PATH environment variable is not set; skipping GCS tests")
-	}
-	gcsClient, err := gcs.NewGCSClient(context.TODO(), pathToGcsCredentials, "")
-	if err != nil {
-		logrus.WithError(err).Fatalf("CRITICAL error getting GCS client with credentials at %s", pathToGcsCredentials)
-	}
-	return gcsClient.Bucket(GcsBucketRoot)
 }
